+++ app/ante/fees.go
@@ -21,8 +21,44 @@ type MinGasPriceDecorator struct {
 	evmKeeper  EVMKeeper
 }
 
+// EthMinGasPriceDecorator will check if the transaction's fee is at least as large
+// as the MinGasPrices param. If fee is too low, decorator returns error and tx
+// is rejected. This applies to both CheckTx and DeliverTx and regardless
+// if London hard fork or fee market params (EIP-1559) are enabled.
+// If fee is high enough, then call next AnteHandler
+type EthMinGasPriceDecorator struct {
+	feesKeeper FeeMarketKeeper
+	evmKeeper  EVMKeeper
+}
+
+// EthMempoolFeeDecorator will check if the transaction's effective fee is at least as large
+// as the local validator's minimum gasFee (defined in validator config).
+// If fee is too low, decorator returns error and tx is rejected from mempool.
+// Note this only applies when ctx.CheckTx = true
+// If fee is high enough or not CheckTx, then call next AnteHandler
+// CONTRACT: Tx must implement FeeTx to use MempoolFeeDecorator
+type EthMempoolFeeDecorator struct {
+	evmKeeper EVMKeeper
+}
+
+// NewMinGasPriceDecorator creates a new MinGasPriceDecorator instance used only for
+// Cosmos transactions.
 func NewMinGasPriceDecorator(fk FeeMarketKeeper, ek EVMKeeper) MinGasPriceDecorator {
 	return MinGasPriceDecorator{feesKeeper: fk, evmKeeper: ek}
+}
+
+// NewEthMinGasPriceDecorator creates a new MinGasPriceDecorator instance used only for
+// Ethereum transactions.
+func NewEthMinGasPriceDecorator(fk FeeMarketKeeper, ek EVMKeeper) EthMinGasPriceDecorator {
+	return EthMinGasPriceDecorator{feesKeeper: fk, evmKeeper: ek}
+}
+
+// NewEthMempoolFeeDecorator creates a new NewEthMempoolFeeDecorator instance used only for
+// Ethereum transactions.
+func NewEthMempoolFeeDecorator(ek EVMKeeper) EthMempoolFeeDecorator {
+	return EthMempoolFeeDecorator{
+		evmKeeper: ek,
+	}
 }
 
 func (mpd MinGasPriceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
@@ -72,20 +108,8 @@ func (mpd MinGasPriceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate 
 	return next(ctx, tx, simulate)
 }
 
-// EthMinGasPriceDecorator will check if the transaction's fee is at least as large
-// as the MinGasPrices param. If fee is too low, decorator returns error and tx
-// is rejected. This applies to both CheckTx and DeliverTx and regardless
-// if London hard fork or fee market params (EIP-1559) are enabled.
-// If fee is high enough, then call next AnteHandler
-type EthMinGasPriceDecorator struct {
-	feesKeeper FeeMarketKeeper
-	evmKeeper  EVMKeeper
-}
-
-func NewEthMinGasPriceDecorator(fk FeeMarketKeeper, ek EVMKeeper) EthMinGasPriceDecorator {
-	return EthMinGasPriceDecorator{feesKeeper: fk, evmKeeper: ek}
-}
-
+// AnteHandle ensures that the that the effective fee from the transaction is greater than the
+// minimum global fee, which is defined by the  MinGasPrice (parameter) * GasLimit (tx argument).
 func (empd EthMinGasPriceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
 	minGasPrice := empd.feesKeeper.GetParams(ctx).MinGasPrice
 
@@ -138,6 +162,47 @@ func (empd EthMinGasPriceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simul
 				errortypes.ErrInsufficientFee,
 				"provided fee < minimum global fee (%d < %d). Please increase the priority tip (for EIP-1559 txs) or the gas prices (for access list or legacy txs)", //nolint:lll
 				fee.TruncateInt().Int64(), requiredFee.TruncateInt().Int64(),
+			)
+		}
+	}
+
+	return next(ctx, tx, simulate)
+}
+
+// AnteHandle ensures that the provided fees meet a minimum threshold for the validator.
+// This check only for local mempool purposes, and thus it is only run on (Re)CheckTx.
+// The logic is also skipped if the London hard fork and EIP-1559 are enabled.
+func (mfd EthMempoolFeeDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
+	if !ctx.IsCheckTx() || simulate {
+		return next(ctx, tx, simulate)
+	}
+	chainCfg := mfd.evmKeeper.GetChainConfig(ctx)
+	ethCfg := chainCfg.EthereumConfig(mfd.evmKeeper.ChainID())
+
+	baseFee := mfd.evmKeeper.GetBaseFee(ctx, ethCfg)
+	// skip check as the London hard fork and EIP-1559 are enabled
+	if baseFee != nil {
+		return next(ctx, tx, simulate)
+	}
+
+	evmDenom := mfd.evmKeeper.GetEVMDenom(ctx)
+	minGasPrice := ctx.MinGasPrices().AmountOf(evmDenom)
+
+	for _, msg := range tx.GetMsgs() {
+		ethMsg, ok := msg.(*evmtypes.MsgEthereumTx)
+		if !ok {
+			return ctx, errorsmod.Wrapf(errortypes.ErrUnknownRequest, "invalid message type %T, expected %T", msg, (*evmtypes.MsgEthereumTx)(nil))
+		}
+
+		fee := sdk.NewDecFromBigInt(ethMsg.GetFee())
+		gasLimit := sdk.NewDecFromBigInt(new(big.Int).SetUint64(ethMsg.GetGas()))
+		requiredFee := minGasPrice.Mul(gasLimit)
+
+		if fee.LT(requiredFee) {
+			return ctx, errorsmod.Wrapf(
+				errortypes.ErrInsufficientFee,
+				"insufficient fees; got: %s required: %s",
+				fee, requiredFee,
 			)
 		}
 	}
