+++ app/ante/fees_test.go
@@ -345,3 +345,7 @@ func (s AnteTestSuite) TestEthMinGasPriceDecorator() {
 		}
 	}
 }
+
+func (suite AnteTestSuite) TestEthMempoolFeeDecorator() {
+	// TODO: add test
+}
